refactor: domain, view 로 리패키징
