+++ src/main/kotlin/blackjack/domain/Dealer.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 class Dealer(
     cards: Cards,
