+++ src/main/kotlin/blackjack/domain/Players.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 class Players(
     private val players: List<Player>,
