+++ src/main/kotlin/blackjack/domain/Suit.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 enum class Suit {
     CLUBS,
