+++ src/test/kotlin/blackjack/domain/CardsTest.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.datatest.withData
